requirements.txt
python-dotenv
requests
python-certifi-win32
openai
psycopg2
Flask